test/run: check DAMON_STAT enabled file existence

+++ tests/run.sh
@@ -6,7 +6,8 @@ cd "$bindir" || exit 1
 
 restart_damon_stat="false"
 damon_stat_enabled_file="/sys/module/damon_stat/parameters/enabled"
-if [ $(cat "$damon_stat_enabled_file") = "Y" ]
+if [ -f "$damon_stat_enabled_file" ] && \
+	[ $(cat "$damon_stat_enabled_file") = "Y" ]
 then
 	echo "DAMON_STAT is running.  Disable for testing."
 	echo N > "$damon_stat_enabled_file"
